Refactoring validations in separate files
